Upgrade: - golang 1.20.4 - bee v2.0.5 - vendor

+++ build/0.pack.sh
@@ -3,7 +3,9 @@
 set -e
 
 PKGFILE=openvpn-web-ui.tar.gz
-rm ../$PKGFILE
+if [ -f $PKGFILE ]; then
+    rm ../$PKGFILE
+fi
 
 time docker run \
     -v "$PWD/../":/go/src/github.com/vuonglequoc/openvpn-web-ui \
